+++ cpp_projs/CPP09/ex01/RPN.hpp
@@ -6,7 +6,7 @@
 #include <stack>
 class RPN{
     private:
-        std::stack<long long> queue;
+        std::stack<long double> queue;
     public:
         RPN();
         ~RPN();
